Updated the event model so that it caches its practice_id upon loading from the db, so that if the event is re-parented to another practice, we can trigger cache updates in both the old and the new practices.
I also played around with dynamically disabling the practice select control in the new event popup.  That's legit syntax (and was a bit tricky) but it turns out to be a nop b/c that partial is only ever rendered for newly-created event records.  Leaving it in in case we ever need to know what that syntax is.